refactor types and file structure for components and theming
