--- src/components/postListItem.tsx
@@ -1,42 +1,0 @@
-import React from 'react';
-import { Column } from './ui/container/column';
-import styled from 'styled-components';
-import { theme } from '../theme/theme';
-import { TextMuted } from './ui/content/textMuted';
-
-interface PostListItemProps {
-  title: string;
-  summary: string;
-  date?: string;
-}
-
-const ItemHeader = styled.h4`
-  font-weight: bold;
-`;
-
-const Item = styled.div`
-  cursor: pointer;
-  border-radius: var(--border-radius-small);
-  transition: all 0.2s ease;
-  &:hover {
-    background: var(--bg-color-highlight);
-    transform: translateY(-2px);
-  }
-  @media (min-width: ${theme.breakpoints.medium}) {
-    padding: var(--margin-small) var(--margin-large);
-  }
-`;
-
-export const PostListItem: React.FC<PostListItemProps> = ({ title, summary, date }) => {
-  return (
-    <Column>
-      <Item>
-        <ItemHeader>{title}</ItemHeader>
-        {date && <TextMuted>{date}</TextMuted>}
-        <p>{summary}</p>
-      </Item>
-    </Column>
-  );
-};
-
-PostListItem.displayName = 'PostListItem';
