+++ src/data/toolBoxCategories.ts
@@ -1,9 +1,6 @@
-interface ToolboxCategory {
-  title: string;
-  items: string[];
-}
+import { ToolboxCategory } from '../types/ToolboxCategory';
 
-export const toolBox: ToolboxCategory[] = [
+export const toolboxCategories: ToolboxCategory[] = [
   {
     title: 'Nära till hands',
     items: ['HTML5, CSS och Preprocessors', 'JavaScript och TypeScript', 'React och Next.js', 'Node', 'Git', 'Jest'],
