--- src/theme/margins.ts
@@ -1,5 +1,0 @@
-export const margins = {
-  small: '5px',
-  medium: '10px',
-  large: '20px',
-};
