+++ src/theme/parts/fonts.ts
@@ -1,22 +1,4 @@
-interface Fonts {
-  fontFamily: {
-    main: string;
-    headings: string;
-    mono: string;
-  };
-  size: {
-    base: string;
-    h1: string;
-    h2: string;
-    h3: string;
-    h4: string;
-    h5: string;
-    h6: string;
-    small: string;
-    smaller: string;
-    code: string;
-  };
-}
+import { Fonts } from '../../types/theme/Fonts';
 
 export const fonts: Fonts = {
   fontFamily: {
