+++ src/theme/parts/margins.ts
@@ -1,0 +1,7 @@
+import { Margins } from '../../types/theme/Margins';
+
+export const margins: Margins = {
+  small: '5px',
+  medium: '10px',
+  large: '20px',
+};
