+++ src/types/Project.ts
@@ -1,0 +1,8 @@
+export interface Project {
+  title: string;
+  icon: string;
+  summary: string;
+  tags: string[];
+  urlSource: string;
+  url?: string;
+}
