+++ src/types/ToolboxCategory.ts
@@ -1,0 +1,4 @@
+export interface ToolboxCategory {
+  title: string;
+  items: string[];
+}
