+++ src/types/theme/Breakpoints.ts
@@ -1,0 +1,5 @@
+export interface Breakpoints {
+  small: string;
+  medium: string;
+  large: string;
+}
