+++ src/types/theme/Margins.ts
@@ -1,0 +1,5 @@
+export interface Margins {
+  small: string;
+  medium: string;
+  large: string;
+}
